Make Tree object structure smaller and faster
The first implementation of the Tree structure was very simplistic: a large
dict with the paths of each directory being the keys, and the tuple of size,
date the values.  This structure has three drawbacks: first, it easily grows
too large; second, it requires a linear search among all branches to find
children values, and third, it requires the values to be sorted each time
they are displayed.

The new structure is a dict of lists.  The dict key is the full path to
parent directory, and the leaves are (possibly ordered) lists of lists,
being the final lists a tripled of (bare) leaf name, size and optional time,
like that (omiting date field):

{
 '': [['/', 3456]],
 '/': [['a/', 2456], ['k/', 932]],
 'a/': [['b/', 12345], ['n/', 23456]],
 'a/b/': [['x/', 123], ['y/', 345]],
 'a/b/y/': [['z/', 67]],
 'k/': [['r/', 876], ['s/', 76]],
}

That is, each path (parent) knows the relative names and values (size and,
optionally, date) of their children as a three value tuple, but not about
itself nor their grandchildren.  If the children have grandchildren, those
children have the full path as name, and the relative names of the children,
keep their data.  Finally, if there are no descendants, there is no specific
node for the child.  The '/' node knows about first-level leaves, and
finally, '' knows about '/'.

On a my tests, the number of entries of the dict is one fourth of the
previous structure (34746 vs 8746), Searching for a full path is a single
dict hit, and the leaves order is preserved.  QED.