Provide better font scaling using float division
Use __future__ module to use float division to provide better font scaling.
Also, when there is a single font size, choose size 2 instead of 5 (was
middle of max and min).  With the old value, font size was too large; as
large as the page title.  For text based values, font size 2 provides good
reading size and enough information on screen.